requirements.txt
flask
flask_sqlalchemy
flask_wtf
wtforms==3.1.2
numpy
transformers
speechrecognition==3.10.1
pyttsx3==2.90
gtts==2.5.1
langchain==0.0.343
pyaudio
sqlalchemy==2.0.30
bcrypt==4.1.2
mysql-connector-python==8.3.0
scikit-learn==1.2.2
jsonify
email-validator
gunicorn
Werkzeug
PyMySQL
python-dotenv